requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
jinja2==3.1.4
langchain==0.3.1
langchain_community==0.3.1
langchain_core==0.3.6
langchain-together==0.2.0
weaviate-client==4.9.5
# torch==2.5.1
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
pandas==2.1.2
transformers==4.46.2
sentence-transformers==3.3.1